requirements.txt
google-genai>=1.0.0
python-docx>=1.1.0
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.0.0
